libs/llm-email/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-email"
version = "0.1.0"
description = "Send emails via macOS Mail.app — compose, draft, send with AppleScript"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "CMDOP Team", email = "team@cmdop.com" }
]
keywords = [
    "email",
    "macos",
    "mail-app",
    "applescript",
    "llm",
    "cmdop",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Typing :: Typed",
]

dependencies = [
    "tortoise-orm>=0.21.0",
    "aiosqlite>=0.20.0",
    "streamlit>=1.40.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
interactive = [
    "click>=8.3.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
email-dashboard = "llm_email.dashboard.app:main"

[project.urls]
Homepage = "https://cmdop.com/skills/llm-email/"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_email"]

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "README.md",
    "pyproject.toml",
    "LICENSE",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
